+++ src/main/java/org/thingai/app/api/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package org.thingai.app.config;
+package org.thingai.app.api.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
